move func to helpers
